requirements.txt
branca==0.4.2
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
distro==1.5.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.22
numpy==1.19.4
openpyxl==3.0.6
pandas==1.2.0
plotly==4.14.3
protobuf==3.14.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.22
tabula-py==2.2.0
urllib3==1.26.2
Werkzeug==1.0.1
xlrd==2.0.1
